+++ CHANGELOG.md
@@ -6,7 +6,8 @@ All notable changes to DocAnvil will be documented in this file.
 
 ### Added
 
-- Added front matter handling for the addition of SEO oriented meta tags in the outputted pages.
+- Added `--path` option to `serve` and `build` commands to target a project from any directory
+- Added front matter handling for the addition of SEO oriented meta tags in the outputted pages
 
 ### Changed